+++ app/src/main/java/org/sopt/and/presentation/auth/signup/component/SignUpTextField.kt
@@ -1,8 +1,7 @@
-package org.sopt.and.presentation.sign.component
+package org.sopt.and.presentation.auth.signup.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -17,36 +16,29 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.sopt.and.R
 import org.sopt.and.core.designsystem.theme.Gray5
 import org.sopt.and.core.designsystem.theme.WavveDisabled
 import org.sopt.and.core.designsystem.theme.White
 
 @Composable
-fun SignUpPasswordField(
+fun SignUpTextField(
     value: String,
     hint: String,
     isValid: Boolean,
     onFocusChange: (Boolean) -> Unit,
-    onValueChange: (String) -> Unit
+    onValueChange: (String) -> Unit,
 ) {
-    var passwordVisible by remember { mutableStateOf(false) }
     var isFocused by remember { mutableStateOf(false) }
-    val borderColor =
-        if (value.isNotEmpty() && !isFocused && !isValid) Color.Magenta else Color.Transparent
+    val borderColor = if (value.isNotEmpty() && !isFocused && !isValid) Color.Magenta else Color.Transparent
     val focusManager = LocalFocusManager.current
     val keyboardController = LocalSoftwareKeyboardController.current
 
@@ -54,7 +46,6 @@ fun SignUpPasswordField(
         value = value,
         onValueChange = onValueChange,
         textStyle = TextStyle(color = White, fontSize = 16.sp),
-        visualTransformation = if (!passwordVisible) PasswordVisualTransformation() else VisualTransformation.None,
         decorationBox = { innerTextField ->
             Box(
                 modifier = Modifier
@@ -65,20 +56,11 @@ fun SignUpPasswordField(
                     Text(text = hint, color = WavveDisabled, fontSize = 16.sp)
                 }
                 innerTextField()
-                Text(
-                    text = if (passwordVisible) stringResource(R.string.sign_up_button_hide) else stringResource(
-                        R.string.sign_up_button_show
-                    ),
-                    color = White,
-                    fontSize = 16.sp,
-                    modifier = Modifier
-                        .align(Alignment.CenterEnd)
-                        .clickable { passwordVisible = !passwordVisible }
-                )
+
             }
         },
         keyboardOptions = KeyboardOptions.Default.copy(
-            imeAction = ImeAction.Done
+            imeAction = ImeAction.Next
         ),
         keyboardActions = KeyboardActions(
             onDone = {
